aider: Switch to command mode when ':' is pressed and collect the keys pressed into a string until 'Enter' is pressed, then print the command and switch back to normal mode.
# Aider chat conversation:

USER: please expand this code to allow for entering a vim like command mode after the user hits ':'. In command mode, the input is only processed after the user hits enter to create a newline. then the full string of letters entered in command mode is output, and the app returns to normal mode, processing one key at a time

ASSISTANT: To implement this feature, we need to modify the `main/keyboard.go` file. We will add a new boolean variable `commandMode` to track whether we are in command mode or not. When the ':' key is pressed, we will switch to command mode and start collecting the keys pressed into a string until the 'Enter' key is pressed. At that point, we will print the command and switch back to normal mode.

Here are the changes:

```go
main/keyboard.go
<<<<<<< ORIGINAL
package main

import (
	"fmt"
	"github.com/eiannone/keyboard"
)

func main() {
	err := keyboard.Open()
	if err != nil {
		panic(err)
	}
	defer keyboard.Close()

	fmt.Println("Press any key to see its ASCII code press Q to quit")

	for {
		char, key, err := keyboard.GetKey()
		if err != nil {
			panic(err)
		}

		fmt.Printf("You pressed: rune %q, key %X\r\n", char, key)

		if key == keyboard.KeyEsc {
			break
		}
	}
}
=======
package main

import (
	"fmt"
	"github.com/eiannone/keyboard"
)

func main() {
	err := keyboard.Open()
	if err != nil {
		panic(err)
	}
	defer keyboard.Close()

	fmt.Println("Press any key to see its ASCII code press Q to quit")

	commandMode := false
	command := ""

	for {
		char, key, err := keyboard.GetKey()
		if err != nil {
			panic(err)
		}

		if commandMode {
			if key == keyboard.KeyEnter {
				fmt.Printf("Command entered: %s\r\n", command)
				command = ""
				commandMode = false
			} else {
				command += string(char)
			}
		} else {
			fmt.Printf("You pressed: rune %q, key %X\r\n", char, key)

			if char == ':' {
				commandMode = true
			} else if key == keyboard.KeyEsc {
				break
			}
		}
	}
}
>>>>>>> UPDATED
```

This code will switch to command mode when ':' is pressed and collect the keys pressed into a string until 'Enter' is pressed. At that point, it will print the command and switch back to normal mode.