aider: Register the new 'shape' sub command with the root command.
# Aider chat conversation:

USER: please also register this new sub command with the root command

ASSISTANT: Sure, I can help with that. However, I need to know the details of the new sub-command that you want to register with the root command. Could you please provide the name of the sub-command and its functionality?

USER: please register the new 'shape' sub command with the root command

ASSISTANT: To register the 'shape' sub-command with the root command, you need to add it to the `init()` function in the `cmd/root.go` file. Here is the change you need to make:

```go
cmd/root.go
<<<<<<< ORIGINAL
func init() {
	RootCmd.AddCommand(NextCmd)
	RootCmd.AddCommand(PrevCmd)
	RootCmd.AddCommand(WalkCmd)
	RootCmd.AddCommand(ParentCmd)
	RootCmd.AddCommand(ChildCmd)
}
=======
func init() {
	RootCmd.AddCommand(NextCmd)
	RootCmd.AddCommand(PrevCmd)
	RootCmd.AddCommand(WalkCmd)
	RootCmd.AddCommand(ParentCmd)
	RootCmd.AddCommand(ChildCmd)
	RootCmd.AddCommand(ShapeCmd)
}
>>>>>>> UPDATED
```

This change adds the `ShapeCmd` to the list of commands that are registered with the root command.