aider: Implemented 'walk' subcommand that walks to the next element for a number of steps.
# Aider chat conversation:

USER: Let's implement a 'walk' subcommand, that takes one arguement, the number of step to take. It then calls the next comamnd internally for the number of steps

ASSISTANT: To implement the 'walk' subcommand, we need to modify the `cmd/navigate.go` file.

Here are the steps:
1. Define a new command `WalkCmd` similar to `NextCmd` and `PrevCmd`.
2. In the `Run` function of `WalkCmd`, parse the argument to get the number of steps.
3. Call the `Run` function of `NextCmd` for the number of steps.

Here is the code change:

```go
cmd/navigate.go
<<<<<<< ORIGINAL
var PrevCmd = &cobra.Command{
	Use:   "prev [selector]",
	Short: "Navigate to the previous element",
	Args:  cobra.MaximumNArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		// TODO: Implement the logic for prev command
	},
}
=======
var WalkCmd = &cobra.Command{
	Use:   "walk [steps]",
	Short: "Walk to the next element for a number of steps",
	Args:  cobra.ExactArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		steps, err := strconv.Atoi(args[0])
		if err != nil {
			fmt.Println("Error: Invalid number of steps.")
			return
		}
		for i := 0; i < steps; i++ {
			NextCmd.Run(cmd, []string{})
		}
	},
}

var PrevCmd = &cobra.Command{
	Use:   "prev [selector]",
	Short: "Navigate to the previous element",
	Args:  cobra.MaximumNArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		// TODO: Implement the logic for prev command
	},
}
>>>>>>> UPDATED
```

Please note that this code assumes that the `NextCmd.Run` function can be called multiple times in a row. If this is not the case, additional changes may be required.